used axios and added fetch call in moviecarasoul itself
