+++ package.json
@@ -18,6 +18,7 @@
     "@react-navigation/elements": "^2.6.2",
     "@react-navigation/native": "^7.1.17",
     "@react-navigation/native-stack": "^7.3.24",
+    "axios": "^1.11.0",
     "react": "19.1.0",
     "react-native": "0.80.2",
     "react-native-safe-area-context": "^5.5.2",
