+++ src/Components/MovieCarasoul/Moviecarasoul.tsx
@@ -1,37 +1,77 @@
+
 /* eslint-disable react-native/no-inline-styles */
-import React from 'react';
-import { FlatList, Image, StyleSheet, Text, View } from 'react-native';
-import { Movie } from '../../types/MovieTypes';
+import React, { useEffect, useState } from 'react';
+import {
+  ActivityIndicator,
+  FlatList,
+  Image,
+  StyleSheet,
+  Text,
+  View,
+} from 'react-native';
+import { MovieApiResponse } from '../../types/MovieTypes';
+import axios from 'axios';
+const options = {
+  method: 'GET',
+  headers: {
+    accept: 'application/json',
+    Authorization:
+      'Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIzZDQ4Njc2MjhiMjEzYTNjNTI5MGZlZWRlZTY5N2UwOSIsIm5iZiI6MTc1MTAxMDExOC42MTY5OTk5LCJzdWIiOiI2ODVlNGI0NmYzNzU2MGMwZjc4MDU4YTYiLCJzY29wZXMiOlsiYXBpX3JlYWQiXSwidmVyc2lvbiI6MX0.AZiFqOvnOhM20RFrzXCYx3ZRVubpiz8jepimaFHD0xY',
+  },
+};
 export default function MovieCarousel({
   title,
-  data,
+  url,
 }: {
   title: string;
-  data: Movie[];
+  url: string;
 }) {
+  const [loading, setLoading] = useState(false);
+  const [data, setData] = useState<MovieApiResponse>();
+  const moviedata = data?.results;
+  useEffect(() => {
+    const fetchMovies = async () => {
+      try {
+        setLoading(true);
+        const res = await axios.get(url, options);
+        setData(res.data);
+      } catch (err) {
+        console.error(err);
+      } finally {
+        setLoading(false);
+      }
+    };
+    fetchMovies();
+  }, []);
   return (
     <View style={{ marginTop: 4, display: 'flex', gap: 8 }}>
       <Text style={{ color: 'white', fontSize: 16, fontWeight: '600' }}>
         {title}
       </Text>
-      <FlatList
-        data={data}
-        horizontal
-        keyExtractor={item => item.id.toString()}
-        renderItem={({ item }) => (
-          <View style={styles.imagecontainer}>
-            <Image
-              source={{
-                uri: `https://image.tmdb.org/t/p/w500${item.poster_path}`,
-              }}
-              style={styles.img}
-            />
-            <Text numberOfLines={1} style={{ color: 'white', maxWidth: 120 }}>
-              {item.title}
-            </Text>
-          </View>
-        )}
-      />
+      {loading ? (
+        <View style={styles.loading}>
+        <ActivityIndicator color={"white"}/>
+        </View>
+      ) : (
+        <FlatList
+          data={moviedata}
+          horizontal
+          keyExtractor={item => item.id.toString()}
+          renderItem={({ item }) => (
+            <View style={styles.imagecontainer}>
+              <Image
+                source={{
+                  uri: `https://image.tmdb.org/t/p/w500${item.poster_path}`,
+                }}
+                style={styles.img}
+              />
+              <Text numberOfLines={1} style={{ color: 'white', maxWidth: 120 }}>
+                {item.title}
+              </Text>
+            </View>
+          )}
+        />
+      )}
     </View>
   );
 }
@@ -50,5 +90,11 @@ const styles = StyleSheet.create({
     gap: 12,
     marginRight: 16,
     marginBottom: 12,
+  },
+   loading: {
+    height: 160,
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
   },
 });
