+++ src/screens/HomeScreen/index.tsx
@@ -1,16 +1,8 @@
 /* eslint-disable react-native/no-inline-styles */
 import { getAuth } from '@react-native-firebase/auth';
-import React, { useEffect, useState } from 'react';
-import {
-  ActivityIndicator,
-  Image,
-  ScrollView,
-  StyleSheet,
-  Text,
-  View,
-} from 'react-native';
+import React from 'react';
+import { Image, ScrollView, StyleSheet, Text, View } from 'react-native';
 import Moviecarasoul from '../../Components/MovieCarasoul/Moviecarasoul';
-import { MovieApiResponse } from '../../types/MovieTypes';
 
 const popularmovieurl =
   'https://api.themoviedb.org/3/movie/popular?language=en-US&page=1';
@@ -18,86 +10,11 @@ const upcomingmoveurl =
   'https://api.themoviedb.org/3/movie/upcoming?language=en-US&page=1';
 const topratedmovieurl =
   'https://api.themoviedb.org/3/movie/top_rated?language=en-US&page=1';
-const options = {
-  method: 'GET',
-  headers: {
-    accept: 'application/json',
-    Authorization:
-      'Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIzZDQ4Njc2MjhiMjEzYTNjNTI5MGZlZWRlZTY5N2UwOSIsIm5iZiI6MTc1MTAxMDExOC42MTY5OTk5LCJzdWIiOiI2ODVlNGI0NmYzNzU2MGMwZjc4MDU4YTYiLCJzY29wZXMiOlsiYXBpX3JlYWQiXSwidmVyc2lvbiI6MX0.AZiFqOvnOhM20RFrzXCYx3ZRVubpiz8jepimaFHD0xY',
-  },
-};
+
 export type MovieData = [];
 export default function HomeScreen() {
   const auth = getAuth();
-  const [popularMoviesData, setPopularMoviesData] =
-    useState<MovieApiResponse>();
-  const [upcomingMoviesData, setUpcomingMoviesData] =
-    useState<MovieApiResponse>();
-  const [topratedMoviesData, setTopratedMoviesData] =
-    useState<MovieApiResponse>();
-  const [loading, setLoading] = useState(false);
   const emailusername = auth.currentUser?.email?.split('@')[0];
-  useEffect(() => {
-    setLoading(true);
-    const fetchallMovies = async () => {
-      try {
-        setLoading(true);
-        const [upcomingRes, popularRes, topratedRes] = await Promise.all([
-          fetch(upcomingmoveurl, options),
-          fetch(popularmovieurl, options),
-          fetch(topratedmovieurl, options),
-        ]);
-        const [upcomingData, popularData, topratedData] = await Promise.all([
-          upcomingRes.json(),
-          popularRes.json(),
-          topratedRes.json(),
-        ]);
-
-        setUpcomingMoviesData(upcomingData);
-        setPopularMoviesData(popularData);
-        setTopratedMoviesData(topratedData);
-      } catch (error) {
-        console.error('Error fetching movies:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-    fetchallMovies();
-    // const fetchMovies = async () => {
-    //   try {
-    //     setLoading(true);
-    //     const res = await fetch(popularmovieurl, options);
-    //     const popularData = await res.json().then();
-    //     setPopularMoviesData(popularData);
-    //   } catch (err) {
-    //     console.error(err);
-    //   } finally {
-    //     setLoading(false);
-    //   }
-    //   try {
-    //     setLoading1(true);
-    //     const res = await fetch(upcomingmoveurl, options);
-    //     const upcomingData = await res.json();
-    //     setUpcomingMoviesData(upcomingData);
-    //   } catch (err) {
-    //     console.error(err);
-    //   } finally {
-    //     setLoading1(false);
-    //   }
-    //   try {
-    //     setLoading2(true);
-    //     const res = await fetch(topratedmovieurl, options);
-    //     const topRatedData = await res.json();
-    //     setTopratedMoviesData(topRatedData);
-    //   } catch (err) {
-    //     console.error(err);
-    //   } finally {
-    //     setLoading2(false);
-    //   }
-    // };
-    // fetchMovies();
-  }, []);
-
   return (
     <ScrollView style={styles.container}>
       <View style={{ display: 'flex', gap: 24, marginBottom: 24 }}>
@@ -149,38 +66,11 @@ export default function HomeScreen() {
           </Text>
         </View>
         <View style={{ display: 'flex', gap: 20 }}>
-          {loading ? (
-            <View style={styles.loading}>
-              <ActivityIndicator color={'white'} />
-            </View>
-          ) : (
-            <Moviecarasoul
-              title="Popular movies"
-              data={popularMoviesData?.results!}
-            />
-          )}
+          <Moviecarasoul url={popularmovieurl} title="Popular movies" />
 
-          {loading ? (
-            <View style={styles.loading}>
-              <ActivityIndicator style={{ height: 160 }} color={'white'} />
-            </View>
-          ) : (
-            <Moviecarasoul
-              title="Upcoming Movies"
-              data={upcomingMoviesData?.results!}
-            />
-          )}
+          <Moviecarasoul url={upcomingmoveurl} title="Upcoming Movies" />
 
-          {loading ? (
-            <View style={styles.loading}>
-              <ActivityIndicator color={'white'} />
-            </View>
-          ) : (
-            <Moviecarasoul
-              title="Top-Rated Movies"
-              data={topratedMoviesData?.results!}
-            />
-          )}
+          <Moviecarasoul url={topratedmovieurl} title="Top-Rated Movies" />
         </View>
       </View>
     </ScrollView>
@@ -192,11 +82,5 @@ const styles = StyleSheet.create({
     padding: 14,
     paddingBottom: 24,
     backgroundColor: '#002335',
-  },
-  loading: {
-    height: 160,
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
   },
 });
